+++ crates/beskid_analysis/src/syntax/expressions/expression.rs
@@ -93,9 +93,11 @@ pub(crate) fn parse_expression(pair: Pair<Rule>) -> Result<Spanned<Expression>, 
         Rule::AssignmentExpression => super::assign_expression::parse_assignment_expression(pair),
         Rule::LogicalOrExpression
         | Rule::LogicalAndExpression
+        | Rule::BitwiseOrExpression
         | Rule::BitwiseAndExpression
         | Rule::EqualityExpression
         | Rule::ComparisonExpression
+        | Rule::ShiftExpression
         | Rule::AdditionExpression
         | Rule::MultiplicationExpression => parse_binary_expression(pair),
         Rule::UnaryExpression => {